requirements.txt
django==4.1
djangorestframework==3.14
psycopg2==2.9.7
celery[redis]==5.2.7
celery_singleton==0.3.1
flower==1.2.0
django-cachalot==2.5.2